+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @enschedule/cli
 
+## 1.1.4
+
+### Patch Changes
+
+- change handlerVersion/handlerId to functionVersion/functionId
+- Updated dependencies
+  - @enschedule/worker-api@1.1.4
+  - @enschedule/types@1.1.3
+
 ## 1.1.3
 
 ### Patch Changes