+++ packages/hub/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @enschedule/hub
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @enschedule/worker-api@1.1.4
+  - @enschedule/pg-driver@1.1.4
+  - @enschedule/worker@1.1.4
+  - @enschedule/dashboard@1.1.5
+  - @enschedule/types@1.1.3
+
 ## 1.0.5
 
 ### Patch Changes