+++ packages/pg-driver/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @enschedule/pg-driver
 
+## 1.1.4
+
+### Patch Changes
+
+- change handlerVersion/handlerId to functionVersion/functionId
+- Updated dependencies
+  - @enschedule/types@1.1.3
+
 ## 1.1.3
 
 ### Patch Changes