+++ packages/worker-cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @enschedule/worker-cli
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @enschedule/worker@1.1.4
+
 ## 0.0.1
 
 ### Patch Changes